ci/circleci: Remove production deployment for now
When a workflow is held for manual approval, the GitHub commit status
gets stuck as in progress until the entire workflow finishes, which
will be the case for a majority of commits.

I guess this encourages almost every commit to be a production deploy,
but for the time being I'll hold off on implementing this.