bot: Fix race condition with how handlers and Dispatch were called
Handlers taking a responses channel and not returning any value are
naturally asynchronous, while bot.Dispatch is designed to be blocking
with its sync.WaitGroup usage, so the handler should be called as a
goroutine while Dispatch is called on the main thread.

This fixes an issue where the main thread would exit and finish the
request before all the other work was done.